WIP detailed Commits builds fixed crashing
